feat: подключи статическое добавление картинок на сервер

+++ controllers/products.js
@@ -19,31 +19,29 @@ module.exports.createProduct = (req, res, next) => {
     type,
     about,
     price,
-    totalPrice,
   } = req.body;
+
   const { name, data } = req.files.image;
   const pathImage = path.join(`${__dirname}/../uploads/products-image/${name}`);
-  fs.writeFileSync(pathImage, data);
+  if (fs.existsSync(pathImage)) {
+    return next(new DublicateKeyError('Изображение уже есть на сервере.'));
+  }
 
-  Product.create({
+  return Product.create({
     mainProduct,
     nameProduct,
     type,
     about,
     price,
-    totalPrice,
     image: {
-      dataImage: fs.readFileSync(pathImage),
-      path: pathImage,
+      path: `http://localhost:3001/image/${name}`,
     },
   })
     .then((product) => {
+      fs.writeFileSync(pathImage, data);
       res.send(product);
     })
     .catch((err) => {
-      if (err.code === 11000) {
-        return next(new DublicateKeyError('Такое изображение уже есть на сервере.'));
-      }
       if (err.message === VALIDATION_ERROR) {
         return next(new BadReqError('Переданы некорректные данные при создании продукта.'), err);
       }
@@ -56,11 +54,10 @@ module.exports.deleteProduct = (req, res, next) => {
   Product.findById(req.params.productId)
     .orFail(new NotFoundError('Продукт не найден'))
     .then((product) => {
-      Product.findByIdAndDelete(product._id.toString())
-        .then(() => {
-          fs.rmSync(product.image.path);
-          res.send({ message: 'Продукт успешно удален.' });
-        });
+      Product.findByIdAndDelete(product._id.toString()).then(() => {
+        fs.rmSync(product.image.path);
+        res.send({ message: 'Продукт успешно удален.' });
+      });
     })
     .catch((err) => {
       if (err.name === CAST_ERROR) {
